add wishlist to allproduct, fix wishlist to no auth user, add handler cart to catalog

+++ pages/api/wishlist.js
@@ -5,23 +5,34 @@ import { WishedProduct } from "../../models/WishedProduct";
 
 export default async function handle(req, res) {
     await mongooseConnect();
-    const { user } = await getServerSession(req, res);
+    const session = await getServerSession(req, res);
+    const user = session?.user || null; // якщо user не існує, присвоїти null
 
     if (req.method === 'POST') {
         const { product } = req.body;
-        const wishedDoc = await WishedProduct.findOne({ userEmail: user.email, product });
-        if (wishedDoc) {
-            await WishedProduct.findByIdAndDelete(wishedDoc._id);
-            res.json({ wishedDoc });
+        if (user) {
+            const wishedDoc = await WishedProduct.findOne({ userEmail: user.email, product });
+            if (wishedDoc) {
+                await WishedProduct.findByIdAndDelete(wishedDoc._id);
+                res.json({ wishedDoc });
+            } else {
+                const newWishedDoc = await WishedProduct.create({ userEmail: user.email, product });
+                res.json({ wishedDoc: newWishedDoc });
+            }
         } else {
-            const newWishedDoc = await WishedProduct.create({ userEmail: user.email, product });
-            res.json({ wishedDoc: newWishedDoc });
+            // Обробка випадку, коли користувач не авторизований
+            res.status(401).json({ error: "Користувач не авторизований" });
         }
     }
 
     if (req.method === 'GET') {
-        res.json(
-            await WishedProduct.find({ userEmail: user.email }).populate('product')
-        );
+        if (user) {
+            res.json(
+                await WishedProduct.find({ userEmail: user.email }).populate('product')
+            );
+        } else {
+            // Обробка випадку, коли користувач не авторизований
+            res.status(401).json({ error: "Користувач не авторизований" });
+        }
     }
 }
